+++ app/openai_whisper/core.py
@@ -56,15 +56,20 @@ def language_detection(audio):
 def write_result(
         result: dict, file: BinaryIO, output: Union[str, None]
 ):
+    options = {
+        'max_line_width': 1000,
+        'max_line_count': 10,
+        'highlight_words': False
+    }
     if output == "srt":
-        WriteSRT(ResultWriter).write_result(result, file=file)
+        WriteSRT(ResultWriter).write_result(result, file=file, options=options)
     elif output == "vtt":
-        WriteVTT(ResultWriter).write_result(result, file=file)
+        WriteVTT(ResultWriter).write_result(result, file=file, options=options)
     elif output == "tsv":
-        WriteTSV(ResultWriter).write_result(result, file=file)
+        WriteTSV(ResultWriter).write_result(result, file=file, options=options)
     elif output == "json":
-        WriteJSON(ResultWriter).write_result(result, file=file)
+        WriteJSON(ResultWriter).write_result(result, file=file, options=options)
     elif output == "txt":
-        WriteTXT(ResultWriter).write_result(result, file=file)
+        WriteTXT(ResultWriter).write_result(result, file=file, options=options)
     else:
         return 'Please select an output method!'